mqtt/requirements.txt
paho-mqtt==1.5.1
psycopg2==2.8.6
